Issues with Notes
I was having a thread hang issue with a lotus notes server that I still
haven't solved.  I changed from javax.mail to the latest Jakarta to
see if that helped but it didn't.  I pulled in a timeout manager from
another project and I'm using that to detect the hung thread, so at
least we'll get a warning something is wrong.

I'm guessing the issue with the notes server is just an implementation
issue on the notes side, since it's an 8.x version from ten years ago.
Nothing I can do about that though.

I added a version banner, and an imapDebug option to enable debugging
at the protocol layer if needed.